+++ package.json
@@ -4,13 +4,9 @@
 	"private": true,
 	"dependencies": {
 		"@date-io/date-fns": "^1.3.13",
-		"@emotion/react": "^11.7.1",
-		"@emotion/styled": "^11.6.0",
 		"@material-ui/core": "^4.11.2",
 		"@material-ui/icons": "^4.11.2",
 		"@material-ui/pickers": "^3.3.10",
-		"@mui/icons-material": "^5.2.5",
-		"@mui/material": "^5.2.6",
 		"@types/jest": "^27.0.2",
 		"@types/lodash": "^4.14.161",
 		"@types/node": "^12.12.54",
