+++ src/App.spec.tsx
@@ -71,83 +71,6 @@ describe('App', () => {
 			expect(mockReactListRef.scrollTo).toHaveBeenCalledWith(0);
 		});
 	});
-
-	it('has filter entries', async () => {
-		const { getByLabelText } = renderPage();
-
-		userEvent.click(getByLabelText('filters'));
-
-		expect(getByLabelText('toggle filter pending')).toBeInTheDocument();
-		expect(getByLabelText('toggle filter complete')).toBeInTheDocument();
-		expect(getByLabelText('toggle filter expired')).toBeInTheDocument();
-	});
-
-	it('checks items by default', async () => {
-		const { getByLabelText } = renderPage();
-
-		userEvent.click(getByLabelText('filters'));
-
-		expect(getByLabelText('pending')).toBeChecked();
-	});
-
-	it('toggles checkmark when entry clicked', async () => {
-		const { getByLabelText, getByText } = renderPage();
-
-		userEvent.click(getByLabelText('filters'));
-		userEvent.click(getByText('pending'));
-
-		expect(getByLabelText('pending')).not.toBeChecked();
-	});
-
-	it('persists checked state when reopening menu', async () => {
-		const { getByLabelText, getByText, baseElement } = renderPage();
-
-		userEvent.click(getByLabelText('filters'));
-		userEvent.click(getByText('pending'));
-
-		const backdrop = baseElement.querySelector('.MuiBackdrop-root');
-
-		if (backdrop === null) throw new Error('No backdrop');
-
-		userEvent.click(backdrop);
-		userEvent.click(getByLabelText('filters'));
-
-		expect(getByLabelText('pending')).not.toBeChecked();
-	});
-
-	it('persists checked state on reload', async () => {
-		const { getByLabelText, getByText, unmount } = renderPage();
-
-		userEvent.click(getByLabelText('filters'));
-		userEvent.click(getByText('pending'));
-
-		unmount();
-
-		const { getByLabelText: getByLabelTextTwo } = renderPage();
-
-		userEvent.click(getByLabelTextTwo('filters'));
-
-		expect(getByLabelTextTwo('pending')).not.toBeChecked();
-	});
-
-	it('filters tasks', async () => {
-		loadNow(new Date('1/1/2020'));
-
-		loadTasksApiData({
-			tasks: [makeTask({ due: '1/1/2020, 11:59 PM', task: 'task 1' })],
-		});
-
-		const { getByLabelText, getByText, queryByText } = renderPage();
-
-		await waitFor(() => {
-			expect(getByText('task 1')).toBeInTheDocument();
-		});
-
-		userEvent.click(getByLabelText('filters'));
-		userEvent.click(getByLabelText('toggle filter pending'));
-
-		expect(queryByText('task 1')).not.toBeInTheDocument();
-	});
 });
 
 // TODO: only highlights task on creation, not on re-load from server
