+++ src/App.tsx
@@ -22,7 +22,6 @@ import NavBar from './components/organisms/NavBar';
 import browser from './lib/Browser';
 import { MuiPickersUtilsProvider } from '@material-ui/pickers';
 import DateFnsUtils from '@date-io/date-fns';
-import { DEFAULT_FILTERS } from './components/molecules/FilterButton';
 
 toast.configure();
 
@@ -47,8 +46,6 @@ function usePageViews(): void {
 
 export function App(): JSX.Element {
 	const [lastToday, setLastToday] = useState<Date>();
-	const [filters, setFilters] = useState<Filters>(DEFAULT_FILTERS);
-
 	const handleTodayClick = () => {
 		setLastToday(browser.getNow());
 	};
@@ -63,7 +60,7 @@ export function App(): JSX.Element {
 		<div className={'page-base'}>
 			<MuiPickersUtilsProvider utils={DateFnsUtils}>
 				<QueryClientProvider client={queryClient}>
-					<NavBar onTodayClick={handleTodayClick} onFilterChange={setFilters} />
+					<NavBar onTodayClick={handleTodayClick} />
 
 					<LoadingIndicator />
 
@@ -101,7 +98,7 @@ export function App(): JSX.Element {
 
 							<Route path={'/'}>
 								<Authenticated>
-									<Tasks lastToday={lastToday} filters={filters} />
+									<Tasks lastToday={lastToday} />
 								</Authenticated>
 							</Route>
 						</Switch>
